adk-chatkit/pyproject.toml
[project]
name = "adk-chatkit"
version = "0.0.3"
description = "OpenAI chatkit backend support for Google ADK"
readme = "README.md"
authors = [
    { name = "Sachdeva, Kapil"}
]
requires-python = ">=3.11"
dependencies = [
    "openai-chatkit>=1.1.0",
    "google-adk>=1.16.0",
    "pydantic>=2.11.7",
]

classifiers = [
  "Intended Audience :: Developers",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "License :: OSI Approved :: Apache Software License",
]

[project.urls]
repository = "https://github.com/ksachdeva/adk-chatkit"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
